[REFACTOR] Move mesh and viewport component files
